+++ training/controllers/__init__.py
@@ -1,1 +1,1 @@
-from .all_controllers import *
+#from .all_controllers import *
