+++ training/templates/bad_form.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+<head>
+	<title>Bienvenido</title>
+</head>
+<body>
+	<h1>Error, Bad form, email or username repeated!</h1>
+</body>
+</html>
